When deleting a database file, it should confirm if the file is a real database file. In case of media provider, it uses Write Ahead Logging mode and there could be wal and shm files for databases. After media scanning, basically 6 files can exist, 3 for internal db, 3 for external db. In that case, unfortunately, a wal file or shm file of a valid database can be deleted during media provider initialization. (DatabaseHelper.onOpen) If a wal file is deleted, db can't guranteee that user data stays safe. This contains the bug fix.

+++ src/com/android/providers/media/MediaProvider.java
@@ -442,7 +442,9 @@ public class MediaProvider extends ContentProvider {
                     long time = other.lastModified();
                     if (time < twoMonthsAgo) {
                         if (LOCAL_LOGV) Log.v(TAG, "Deleting old database " + databases[i]);
-                        mContext.deleteDatabase(databases[i]);
+                        if(databases[i].endsWith(".db")) {
+                            mContext.deleteDatabase(databases[i]);
+                        }
                         databases[i] = null;
                         count--;
                     }
@@ -468,7 +470,9 @@ public class MediaProvider extends ContentProvider {
                 // delete least recently used database
                 if (lruIndex != -1) {
                     if (LOCAL_LOGV) Log.v(TAG, "Deleting old database " + databases[lruIndex]);
-                    mContext.deleteDatabase(databases[lruIndex]);
+                    if(databases[lruIndex].endsWith(".db")) {
+                        mContext.deleteDatabase(databases[lruIndex]);
+                    }
                     databases[lruIndex] = null;
                     count--;
                 }
